Other Models Code/requirements.txt
amqp==5.1.1
anyio==3.6.2
apache-libcloud==3.6.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arpy==2.3.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==6.0.0
brew==0.1.3
bs4==0.0.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
comm==0.1.3
commonmark==0.9.1
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
depot==0.0.12
depot-pm==0.1.11
distlib==0.3.5
distro==1.8.0
docopt==0.6.2
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.8.0
Flask==2.2.2
flask-sock==0.5.2
Flask-SQLAlchemy==2.5.1
fonttools==4.37.1
fqdn==1.5.1
gitdb==4.0.9
GitPython==1.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jisho-api==0.1.8
joblib==1.2.0
joe==0.0.7
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
kombu==5.2.4
lockfile==0.12.2
lxml==4.9.2
MarkupSafe==2.1.1
marshmallow==3.19.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.5
mplleaflet==0.0.5
mpmath==1.3.0
multipledispatch==0.6.0
nano==0.10.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.8.6
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
packaging==21.3
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pdfminer==20191125
pdfminer.six==20220524
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
pthflops==0.4.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyagrum==1.7.1
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.4
pydot==1.4.2
pygame==2.3.0
Pygments==2.14.0
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.6
pyparsing==3.0.9
pyrr==0.10.3
pyrsistent==0.19.3
pytest==7.2.0
python-dateutil==2.8.2
python-gnupg==0.5.0
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.1
Random-Word==1.0.11
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==10.16.2
romkan==0.2.1
scikit-build==0.17.5
scikit-learn==1.2.2
scipy==1.11.0
seaborn==0.12.2
Send2Trash==1.8.0
simple-websocket==0.8.1
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
stack-data==0.6.2
sympy==1.11.1
taskr==0.2.45
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.13
vine==5.0.0
virtualenv==20.16.3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.7
wonderwords==2.2.0
wsproto==1.2.0
